Added the code for selecting all the person data from a database.
Added a file for the database tests. I still need to make test files for all other modules. Then these need to be filled with unit tests ensuring the functionality of the respective module.

Docstring, Doctest, Help screens and graphs still need to be implemented.

It is also advisable for comments to be written explaining the use of the code. This will make the refactoring taking place by another team in the next assignment a ton easier :^)